temp dot file bug #110
